[IR] Added `DSCF` transformation.
Eliminates local control flow and mutable state.

Converts an input with while loops and local variables (which is already in
ANF) in the so-called `direct-style` representation.

For a description of the term `direct-style control flow` see Section 6.1.3
in [1].

Together with the `ANF` transformation constitutes the bulk of the `Source`
to `Core` translation.

[1] Static Single Assignment Book, Inria
    http://ssabook.gforge.inria.fr/latest/book.pdf